refactor(web): type, constant 분리
