+++ apps/web/src/app/product/[id]/reservation-form/hooks/index.ts
@@ -1,15 +1,2 @@
-export {
-  MAXIMUM_DURATION_HOURS,
-  MAXIMUM_PEOPLE_COUNT,
-  MINIMUM_DURATION_HOURS,
-  MINIMUM_PEOPLE_COUNT,
-  PRIMARY_SCHEDULE_CHOICE_KEY,
-  REQUEST_CONTENT_MAX_LENGTH,
-  SCHEDULE_CHOICE_KEYS,
-  SCHEDULE_CHOICES,
-  hasCompletedSchedule,
-  hasSelectableScheduleChoice,
-} from './reservationCopyFormShared';
-export type { ReservationApplicant } from './reservationCopyFormShared';
 export { default as useReservationCopyForm } from './useReservationCopyForm';
 export type { ReservationCopyFormModel } from './useReservationCopyForm';
