+++ apps/web/src/app/product/[id]/reservation-form/types/copy.ts
@@ -1,0 +1,18 @@
+import { SCHEDULE_CHOICES } from '../constants/copy-form';
+
+export type ReservationApplicant = {
+  name: string;
+  phoneNumber: string;
+  email: string;
+};
+
+export type ScheduleChoiceKey = (typeof SCHEDULE_CHOICES)[number]['key'];
+
+export type UploadConsentStatus = '' | 'agree' | 'disagree';
+
+export type ScheduleSelectionValue = {
+  date: string;
+  time: string;
+};
+
+export type ScheduleSelections = Record<ScheduleChoiceKey, ScheduleSelectionValue>;
